+++ index.html
@@ -51,7 +51,7 @@
 
     </div>
     
-    <script src="scripts/main.js"></script>
+    <script src="scripts/main.js" type="module"></script>
     
 </body>
 </html>
